+++ pytorch_lightning/trainer/trainer.py
@@ -1723,11 +1723,10 @@ class Trainer(
 
     @property
     def resume_from_checkpoint(self) -> Optional[Union[str, Path]]:
-        """
-        .. deprecated:: v1.5
-            This method is deprecated in v1.5 and will be removed in v1.7.
-            Please use `Trainer.resume_checkpoint_path` instead.
-        """
+        rank_zero_deprecation(
+            "`trainer.resume_from_checkpoint` is deprecated in v1.5 and will be removed in v1.7."
+            " Use `trainer.resume_checkpoint_path` instead."
+        )
         return self.resume_checkpoint_path
 
     @property