+++ tests/loops/optimization/test_optimizer_loop.py
@@ -223,7 +223,6 @@ def test_loop_restart_progress_multiple_optimizers(tmpdir, n_optimizers, stop_op
     model.training_epoch_end = None
     model.optimizer_step = Mock(wraps=model.optimizer_step)
     trainer = Trainer(
-        resume_from_checkpoint=str(tmpdir / ".pl_auto_save.ckpt"),
         default_root_dir=tmpdir,
         max_epochs=n_epochs,
         limit_train_batches=n_batches,
@@ -232,7 +231,7 @@ def test_loop_restart_progress_multiple_optimizers(tmpdir, n_optimizers, stop_op
         logger=False,
         checkpoint_callback=False,
     )
-    trainer.fit(model)
+    trainer.fit(model, ckpt_path=str(tmpdir / ".pl_auto_save.ckpt"))
     weights_resumed = model.parameters()
 
     # check that the final weights of a resumed run match the weights of a run that never failed