+++ tests/plugins/test_sharded_plugin.py
@@ -132,7 +132,7 @@ def test_ddp_sharded_plugin_finetune(tmpdir):
 
 
 @RunIf(skip_windows=True, fairscale=True)
-def test_ddp_sharded_plugin_resume_from_checkpoint(tmpdir):
+def test_ddp_sharded_plugin_fit_ckpt_path(tmpdir):
     """Test to ensure that resuming from checkpoint works."""
     model = BoringModel()
     trainer = Trainer(accelerator="ddp_sharded_spawn", num_processes=2, fast_dev_run=True)
@@ -145,16 +145,16 @@ def test_ddp_sharded_plugin_resume_from_checkpoint(tmpdir):
     model = BoringModel()
 
     trainer = Trainer(
-        accelerator="ddp_sharded_spawn", num_processes=2, fast_dev_run=True, resume_from_checkpoint=checkpoint_path
+        accelerator="ddp_sharded_spawn", num_processes=2, fast_dev_run=True
     )
 
-    trainer.fit(model)
+    trainer.fit(model, ckpt_path=checkpoint_path)
 
 
 @pytest.mark.skip(reason="Not a critical test, skip till drone CI performance improves.")  # todo
 @pytest.mark.skip(reason="Currently unsupported restarting training on different number of devices.")
 @RunIf(min_gpus=2, skip_windows=True, fairscale=True)
-def test_ddp_sharded_plugin_resume_from_checkpoint_downsize_gpus(tmpdir):
+def test_ddp_sharded_plugin_fit_ckpt_path_downsize_gpus(tmpdir):
     """Test to ensure that resuming from checkpoint works when downsizing number of GPUS."""
     model = BoringModel()
     trainer = Trainer(accelerator="ddp_sharded_spawn", fast_dev_run=True, gpus=2)
@@ -167,14 +167,14 @@ def test_ddp_sharded_plugin_resume_from_checkpoint_downsize_gpus(tmpdir):
     model = BoringModel()
 
     trainer = Trainer(
-        accelerator="ddp_sharded_spawn", fast_dev_run=True, gpus=1, resume_from_checkpoint=checkpoint_path
+        accelerator="ddp_sharded_spawn", fast_dev_run=True, gpus=1
     )
 
-    trainer.fit(model)
+    trainer.fit(model, ckpt_path=checkpoint_path)
 
 
 @RunIf(min_gpus=1, skip_windows=True, fairscale=True)
-def test_ddp_sharded_plugin_resume_from_checkpoint_gpu_to_cpu(tmpdir):
+def test_ddp_sharded_plugin_fit_ckpt_path_gpu_to_cpu(tmpdir):
     """Test to ensure that resuming from checkpoint works when going from GPUs- > CPU."""
     model = BoringModel()
     trainer = Trainer(accelerator="ddp_sharded_spawn", gpus=1, fast_dev_run=True)
@@ -187,10 +187,10 @@ def test_ddp_sharded_plugin_resume_from_checkpoint_gpu_to_cpu(tmpdir):
     model = BoringModel()
 
     trainer = Trainer(
-        accelerator="ddp_sharded_spawn", num_processes=2, fast_dev_run=True, resume_from_checkpoint=checkpoint_path
+        accelerator="ddp_sharded_spawn", num_processes=2, fast_dev_run=True
     )
 
-    trainer.fit(model)
+    trainer.fit(model, ckpt_path=checkpoint_path)
 
 
 @RunIf(skip_windows=True, special=True, fairscale=True)